chore: Refactoring focus strategy to get the correct key name on history removal for all IDEs (#39689)
## Description

Refactoring focus strategy to get the correct key name on history
removal for all IDEs

Fixes [#39597](https://github.com/appsmithorg/appsmith/issues/39597)

## Automation

/ok-to-test tags="@tag.All"

### :mag: Cypress test results
