+++ app/client/src/ce/navigation/FocusElements/AppIDE.ts
@@ -55,10 +55,10 @@ import {
   setSelectedEntityUrl,
   setSelectedJSObject,
   setSelectedQuery,
-} from "ee/navigation/FocusSetters";
+} from "ee/navigation/AppIDEFocusSetters";
 import { getFirstDatasourceId } from "selectors/datasourceSelectors";
 import { FocusElement, FocusElementConfigType } from "navigation/FocusElements";
-import type { FocusElementsConfigList } from "sagas/FocusRetentionSaga";
+import type { FocusElementsConfigList } from "ee/navigation/FocusStrategy/types";
 import { ActionExecutionResizerHeight } from "PluginActionEditor/components/PluginActionResponse/constants";
 import {
   getPluginActionConfigSelectedTab,
