+++ app/client/src/ce/navigation/FocusStrategy/NoIDEFocusStrategy.ts
@@ -1,4 +1,4 @@
-import type { FocusStrategy } from "sagas/FocusRetentionSaga";
+import type { FocusStrategy } from "ee/navigation/FocusStrategy/types";
 import NoIDEFocusElements from "../FocusElements/NoIDE";
 export const NoIDEFocusStrategy: FocusStrategy = {
   focusElements: NoIDEFocusElements,
@@ -8,8 +8,11 @@ export const NoIDEFocusStrategy: FocusStrategy = {
   *getEntitiesForStore() {
     return [];
   },
-  getEntityParentUrl(): string {
+  getEntityParentUrl: (): string => {
     return "";
+  },
+  getUrlKey: function* (url: string) {
+    return url;
   },
   *waitForPathLoad() {},
 };
