+++ app/client/src/ee/navigation/AppIDEFocusSetters.ts
@@ -1,0 +1,1 @@
+export * from "ce/navigation/AppIDEFocusSetters";
