--- app/client/src/ee/navigation/FocusSetters.ts
@@ -1,1 +1,0 @@
-export * from "ce/navigation/FocusSetters";
