+++ app/client/src/ee/navigation/FocusStrategy/types.ts
@@ -1,0 +1,1 @@
+export * from "ce/navigation/FocusStrategy/types";
